Change <CR> behavior for coc.nvim.
Perhaps the recommended settings I'd copied are outdated; <CR> as it was
didn't select a completion. I thought this was due to a collision with
the auto-pairs plugin, but turning it off and coc-pairs on didn't help.
The coc-pairs docs talk about the <CR> mapping I've replaced the old one
with, and now things Just Work™.